requirements.txt
streamlit
PyMuPDF
pytesseract
pdf2image
gTTS
python-dotenv
google-generativeai
Pillow
e.